Fixed memory leak
